+++ onban/Important/SceneDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  SceneDelegate.swift
+//  onban
+//
+//  Created by Zeto on 2022/10/27.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
+
